Add files via upload
This one can be used for one part of the submission. In the notebook, the text extraction from json files are included, however not used in the final query due to memory limitation. But how the merge the text and meta data are shown. It used ranking from BM25, which is ranking function which calculates score to represent a document's relevance with respect to query. In tests this approach gives better results compared to earlier TF-IDF based scoring.